moved queryParamNames and useReplaceQueryParameter to common
