+++ src/common/useReplaceQueryParameter.js
@@ -1,5 +1,5 @@
 import { useHistory, useLocation } from "react-router-dom";
-import { paginationQueryParamName, searchQueryParamName } from "../../queryParamNames";
+import { paginationQueryParamName, searchQueryParamName } from "./queryParamNames";
 
 export const useReplaceQueryParameter = () => {
     const location = useLocation();
@@ -17,7 +17,6 @@ export const useReplaceQueryParameter = () => {
             searchParams.delete(paginationQueryParamName);
         }
         
-
         history.push(`${location.pathname.includes("/movie") ? "/movies" : "/people"}?${searchParams.toString()}`);
     };
 };
